+++ packages/yourloops/test/integration/pages/profile/profile.patient.spec.tsx
@@ -106,6 +106,7 @@ describe('Profile page for patient', () => {
     expect(fields.genderSelect).toHaveTextContent('Male')
     expect(fields.hba1cInput).toHaveValue('7.5%')
     expect(saveButton).toBeDisabled()
+    expect(screen.queryByTestId('country-selector')).not.toBeInTheDocument()
 
     fireEvent.mouseDown(within(screen.getByTestId('profile-local-selector')).getByRole('combobox'))
     fireEvent.click(screen.getByRole('option', { name: 'English' }))