+++ CHANGELOG.md
@@ -8,6 +8,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2850 Add product name (model) in pump information card
 - YLP-2851 Remove pump cartridge expiration date from Device view and PDF report
 - YLP-2859 Display alarms from Dana-i5 pump
+- YLP-2863 As an HCP, I want to be able to change my country in my profile settings
 - YLP-2868 As a HCP/caregiver, I want to see my patient listed by alphabetical order based on the last name
 
 ### Fixed